Fix dynamically-loaded config
