+++ src/js/App/RootApp.js
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { safeLoad } from 'js-yaml';
 import { connect } from 'react-redux';
 import { useScalprum, ScalprumRoute, ScalprumLink } from '@scalprum/react-core';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
@@ -9,17 +8,10 @@ import analytics from '../analytics';
 import sentry from '../sentry';
 import createChromeInstance from '../chrome/create-chrome';
 import registerUrlObserver from '../url-observer';
-import sourceOfTruth from '../nav/sourceOfTruth';
 
-const RootApp = () => {
+const RootApp = ({ config }) => {
+  const scalprum = useScalprum(config);
   const [insights, setInsights] = useState();
-  const [config, setConfig] = useState();
-
-  console.log('About to use Scalprum with this value:');
-  console.log(config);
-  let scalprum = useScalprum(config);
-  console.log('Scalprum initialized:');
-  console.log(scalprum);
 
   useEffect(() => {
     const libjwt = auth();
@@ -40,28 +32,12 @@ const RootApp = () => {
     const insights = window.insights;
     setInsights(insights);
 
-    sourceOfTruth('testPrefix')
-      .then((configYaml) => {
-        let appConfig = safeLoad(configYaml);
-        Object.entries(appConfig).forEach(([key, val]) => {
-          val['scriptLocation'] = `${window.location.origin}${val['scriptLocation']}`;
-        });
-        console.log('Config is done:');
-        console.log(appConfig);
-        return appConfig;
-      })
-      .then((appConfig) => {
-        console.log('Setting config:');
-        console.log(appConfig);
-        setConfig(appConfig);
-      });
-
     if (typeof _satellite !== 'undefined' && typeof window._satellite.pageBottom === 'function') {
       window._satellite.pageBottom();
       registerUrlObserver(window._satellite.pageBottom);
     }
   }, []);
-  if (!scalprum || !scalprum.initialized || !insights) {
+  if (!scalprum.initialized || !insights) {
     return (
       <div>
         <h1>Loading</h1>
@@ -102,7 +78,8 @@ RootApp.propTypes = {
   activeLocation: PropTypes.string,
   pageAction: PropTypes.string,
   pageObjectId: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  globalFilterRemoved: PropTypes.bool,
+  globalFilterHidden: PropTypes.bool,
+  config: PropTypes.any,
 };
 
 function stateToProps({ chrome: { activeApp, activeLocation, appId, pageAction, pageObjectId }, globalFilter: { globalFilterRemoved } = {} }) {
